wip: reorg kzg for eval form variant
